# API client: make http reads more efficient